Removed self test code from sr_subscribe.
